implement email service timeouts. fixes #1051
